Remove NOLINT and convert directly

+++ ttnn/cpp/ttnn/operations/moreh/moreh_group_norm/device/moreh_group_norm_program_factory.cpp
@@ -76,10 +76,7 @@ MorehGroupNormOperation::MorehGroupNormFactory::cached_program_t MorehGroupNormO
         num_groups);
     const auto num_inner_tiles = (num_channels / num_groups) * Ht * Wt;
 
-    // num_channels is guaranteed to be divisible by num_groups (group norm requirement),
-    // so integer division is exact and the cast to float preserves the correct value.
-    // NOLINTNEXTLINE(bugprone-integer-division)
-    const auto f_c = static_cast<float>(num_channels / num_groups);
+    const auto f_c = static_cast<float>(num_channels) / static_cast<float>(num_groups);
     const auto f_ht = static_cast<float>(origin_h) / static_cast<float>(TILE_HEIGHT);
     const auto f_wt = static_cast<float>(origin_w) / static_cast<float>(TILE_WIDTH);
     auto scaler = 1.0f / (static_cast<float>(TILE_WIDTH) * sqrt(f_c * f_ht * f_wt));
